oplog: inline large-file ignore helper
