+++ crates/gitbutler-cherry-pick/Cargo.toml
@@ -17,7 +17,6 @@ gitbutler-commit.workspace = true
 git2.workspace = true
 gix.workspace = true
 anyhow.workspace = true
-tracing.workspace = true
 
 [lints]
 workspace = true
