+++ crates/gitbutler-cherry-pick/src/lib.rs
@@ -4,9 +4,6 @@ use anyhow::{Context as _, Result};
 use but_oxidize::{ObjectIdExt as _, OidExt, RepoExt};
 use gitbutler_commit::commit_ext::CommitExt;
 
-mod repository_ext;
-pub use repository_ext::RepositoryExtLite;
-
 #[derive(Default)]
 pub enum ConflictedTreeKey {
     /// The commit we're rebasing onto "head"
